Messages, GUI, claim roles and save data overhaul
This is a huge commit that primarily includes four major changes:
- All messages have been rewritten to use message-api, which allows for
greater customization of all messages
- The gui has been rewritten with cleaner code and more support for all
available features
- Roles can now be defined per claim. Default roles such as trusted and
moderator exist and can also be modified. This allows claim owners to
give the same set of permissions to many people with just a few clicks
- The loading and saving system has been rewritten to use Codecs. This
change also introduces the use of DataFixerUppers, which handle any
changes made to the claim data format to ensure old data can be loaded.